fix: move hooks to the same folder as the route component. refs #34 refs #37. Fixing conflicts to merge.
