+++ app/src/App.tsx
@@ -14,7 +14,7 @@ function App() {
           </Route>
           <Route path="/">
             <Suspense fallback="loading">
-              <ChoiceBoard/>
+              <ChoiceBoard />
             </Suspense>
           </Route>
         </Switch>
